test fixes/workarounds for Rubinius (see #576)
(1) There is still an issue with Rails 4.1.8, which seems to ignore
config/environments/test.rb and config/initializers/*.rb files.

For now, this test is allowed to fail.

(2) Since the test suite runs in a Docker container, we cannot
install libyaml, which is needed for psych, which in turn is
needed as YAML engine, because Rubinius ships the abandoned
Syck engine...

For now, the two test cases are disabled on Travis, but will run
locally.